Updated to include stat modifiers

+++ Models/BattleStats.cs
@@ -2,7 +2,7 @@
 {
     public class BattleStats
     {
-        public int Evasion { get; set; } = 100;
-        public int Accuracy { get; set; } = 100;
+        public double Evasion { get; set; } = 100;
+        public double Accuracy { get; set; } = 100;
     }
 }
