+++ Models/StatsModel.cs
@@ -2,13 +2,13 @@
 {
     public class StatsModel
     {
-        public int HP { get; set; }
-        public int Attack { get; set; }
-        public int Defense { get; set; }
-        public int SpecialAttack { get; set; }
-        public int SpecialDefense { get; set; }
-        public int Speed { get; set; }
-        public int Total { get; set; }
+        public double HP { get; set; }
+        public double Attack { get; set; }
+        public double Defense { get; set; }
+        public double SpecialAttack { get; set; }
+        public double SpecialDefense { get; set; }
+        public double Speed { get; set; }
+        public double Total { get; set; }
 
         public BattleStats BattleStats = new BattleStats();
     }
